Fix bug 1026639 - Allow deleting attachments.
This moves attachment revision files into a trash.
There is a management command to purge the trash after a number of days that should be run regularly via cron.
This again extends the attachments admin UI somewhat to support this new feature.
It's not offered to the front-end users at all.